Finished 04-chaining-selectors exercise
I practiced application of chaining selectors using it to proportion and distort an <img> element. the proportioned image retained its original shape as width value is fixed and the height set to  auto.
while the distorted image changed shape as both the with and height were set to a fixed value.